requirements.txt

Flask==3.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-SocketIO
python-socketio
python-dotenv==1.0.1
PyMySQL==1.1.0
pandas
openpyxl
reportlab==4.2.0
APScheduler==3.10.4
qrcode[pil]==7.4.2
Flask-Caching==2.1.0
Flask-Migrate==4.0.5
Flask-Compress==1.13
